Add delete notes feature
