+++ NotesApp/Controllers/HomeController.cs
@@ -1,49 +1,12 @@
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using NotesApp.Data;
-using NotesApp.Models;
-using System.Diagnostics;
-using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
 
 namespace NotesApp.Controllers
 {
     public class HomeController : Controller
     {
-        private readonly ILogger<HomeController> _logger;
-        private readonly ApplicationDbContext _context;
-
-        public HomeController(ILogger<HomeController> logger, ApplicationDbContext context)
+        public IActionResult Index()
         {
-            _logger = logger;
-            _context = context;
-        }
-
-        // GET: /Home/Index
-        public async Task<IActionResult> Index()
-        {
-            var notes = await _context.Notes.ToListAsync();
-            return View(notes);
-        }
-
-        // GET: /Home/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: /Home/Create
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create(Note note)
-        {
-            if (ModelState.IsValid)
-            {
-                _context.Notes.Add(note);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(note);
+            return View(); 
         }
 
         public IActionResult Privacy()
@@ -51,13 +14,9 @@ namespace NotesApp.Controllers
             return View();
         }
 
-        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
         public IActionResult Error()
         {
-            return View(new ErrorViewModel 
-            { 
-                RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier 
-            });
+            return View();
         }
     }
 }
