+++ NotesApp/Controllers/NotesController.cs
@@ -1,0 +1,55 @@
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using NotesApp.Data;
+using NotesApp.Models;
+using System.Threading.Tasks;
+
+namespace NotesApp.Controllers
+{
+    public class NotesController : Controller
+    {
+        private readonly ApplicationDbContext _context;
+
+        public NotesController(ApplicationDbContext context)
+        {
+            _context = context;
+        }
+
+        public async Task<IActionResult> Index()
+        {
+            var notes = await _context.Notes.ToListAsync();
+            return View("~/Views/Home/Index.cshtml", notes);
+        }
+
+        public IActionResult Create()
+        {
+            return View("~/Views/Home/Create.cshtml");
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Create(Note note)
+        {
+            if (ModelState.IsValid)
+            {
+                _context.Add(note);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return View("~/Views/Home/Create.cshtml", note);
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Delete(int id)
+        {
+            var note = await _context.Notes.FindAsync(id);
+            if (note != null)
+            {
+                _context.Notes.Remove(note);
+                await _context.SaveChangesAsync();
+            }
+            return RedirectToAction(nameof(Index));
+        }
+    }
+}
