+++ NotesApp/Views/Home/Create.cshtml
@@ -1,24 +1,28 @@
 @model NotesApp.Models.Note
 
-<h2>Create Note</h2>
+@{
+    ViewData["Title"] = "Create Note";
+}
+
+<h2>Create a new Note</h2>
 
 <form asp-action="Create" method="post">
-    <div>
+    <div class="form-group">
         <label asp-for="Title"></label>
-        <input asp-for="Title" />
+        <input asp-for="Title" class="form-control" />
         <span asp-validation-for="Title" class="text-danger"></span>
     </div>
-    <div>
+    <div class="form-group">
         <label asp-for="Content"></label>
-        <textarea asp-for="Content"></textarea>
+        <textarea asp-for="Content" class="form-control"></textarea>
     </div>
-    <div>
+    <div class="form-group">
         <label asp-for="ReminderAt"></label>
-        <input asp-for="ReminderAt" type="datetime-local" />
+        <input asp-for="ReminderAt" type="datetime-local" class="form-control" />
     </div>
-    <button type="submit">Save</button>
+    <button type="submit" class="btn btn-primary">Create</button>
 </form>
 
 @section Scripts {
-    <partial name="_ValidationScriptsPartial" />
+    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
 }
